+++ lgrez/bdd/base.py
@@ -345,7 +345,7 @@ def autodoc_Column(*args, doc="", comment=None, **kwargs):
     sa_type = (f":class:`{col.type!r} "
                f"<sqlalchemy.types.{type(col.type).__name__}>`")
     py_type = col.type.python_type
-    if py_type.__module__ in ("builtins", "lgrez.bdd.model"):
+    if py_type.__module__ in ("builtins", "lgrez.bdd.enums"):
         py_type_str = py_type.__name__
     else:
         py_type_str = f"{py_type.__module__}.{py_type.__name__}"
@@ -363,7 +363,7 @@ def autodoc_Column(*args, doc="", comment=None, **kwargs):
     return col
 
 
-def autodoc_ManyToOne(tablename, *args, doc="", **kwargs):
+def autodoc_ManyToOne(tablename, *args, doc="", nullable=False, **kwargs):
     """Returns Python-side well documented many-to-one relationship.
 
     Represents a relationship where each element in this table is
@@ -373,18 +373,25 @@ def autodoc_ManyToOne(tablename, *args, doc="", **kwargs):
     Example: ``Book.editor`` (each book has an editor, an editor
     publishes several books)
 
-    Use exactly as :func:`sqlalchemy.orm.relationship`.
+    Use exactly as :func:`sqlalchemy.orm.relationship`, plus the
+    keyword argument :attr:`nullable` (see below).
 
     Args:
         \*args, \*\*kwargs: passed to :class:`sqlalchemy.orm.relationship`.
         doc (str): relationship description, enhanced with class name
             and relationship type.
+        nullable (bool): indicates whether the relationship can be
+            omitted (impacts docs only, default ``False``).
 
     Returns:
         :class:`sqlalchemy.orm.RelationshipProperty`
     """
     first, sep, rest = doc.partition("\n")
-    doc = f"~bdd.{tablename}: {first} *(many-to-one relationship)*"
+    or_none = " | ``None``" if nullable else ""
+    doc = (f":class:`~bdd.{tablename}`{or_none}: {first} "
+           "*(many-to-one relationship)*")
+    if not nullable:
+        doc += " (NOT NULL)"
     if rest:
         doc += sep + rest
 
@@ -417,6 +424,38 @@ def autodoc_OneToMany(tablename, *args, doc="", **kwargs):
         doc += sep + rest
 
     return sqlalchemy.orm.relationship(tablename, *args, doc=doc, **kwargs)
+
+
+def autodoc_DynamicOneToMany(tablename, *args, doc="", **kwargs):
+    """Returns Python-side well documented dynamic one-to-many relationship.
+
+    Represents a relationship where each element in this table is
+    linked to **a lot of elements** in ``tablename``, itself back-
+    refering to **one element** in this table.
+
+    The difference with :func:`~.autodoc_OneToMany` is that items are NOT
+    accessible directly through class attribute, which returns a query
+    that but must be fetched first:
+
+    Examples: ``Editor.books.all()``, ``Editor.books.filter_by(...).one()``
+
+    Use exactly as :func:`sqlalchemy.orm.dynamic_loader`.
+
+    Args:
+        \*args, \*\*kwargs: passed to :class:`sqlalchemy.orm.dynamic_loader`.
+        doc (str): relationship description, enhanced with class name
+            and relationship type.
+
+    Returns:
+        :class:`sqlalchemy.orm.RelationshipProperty`
+    """
+    first, sep, rest = doc.partition("\n")
+    doc = (f"~sqlalchemy.orm.query.Query[~bdd.{tablename}]: {first} "
+           "*(dynamic one-to-many relationship)*")
+    if rest:
+        doc += sep + rest
+
+    return sqlalchemy.orm.dynamic_loader(tablename, *args, doc=doc, **kwargs)
 
 
 def autodoc_ManyToMany(tablename, *args, doc="", **kwargs):